Split the directories for source files and header files. Also split the logic so now, everything assembler related resides in assembler.cpp, while everything file I/O related remains in main.cpp
