+++ .github/workflows/Compilation_check.yml
@@ -19,7 +19,7 @@ jobs:
               run: sudo apt-get install -y build-essential
 
             - name: Simple Compile with g++
-              run: g++ -static-libstdc++ ./'New Code'/main.cpp -o Assembler
+              run: g++ -static-libstdc++ ./src/main.cpp -o Assembler
         
             - name: Ensure Executable Works
               run: chmod +x Assembler
@@ -49,7 +49,7 @@ jobs:
                 echo "C:\\mingw64\\bin" >> $env:GITHUB_PATH
                 
             - name: Simple Compile with g++
-              run: g++ -static-libstdc++ ./'New Code'/main.cpp Assembler.exe
+              run: g++ -static-libstdc++ ./src/main.cpp Assembler.exe
 
     build-macos:
         runs-on: macos-latest
@@ -58,7 +58,7 @@ jobs:
               uses: actions/checkout@v3
 
             - name: Simple Compile with clang++
-              run: clang++ -std=c++17 ./'New Code'/main.cpp -o Assembler
+              run: clang++ -std=c++17 ./src/main.cpp -o Assembler
 
             - name: Ensure Executable Works
               run: chmod +x Assembler
